refactor: Tab 식별자 수정
