--- src/pages/admin/NoticeAdminTab.tsx
@@ -1,5 +1,0 @@
-const NoticeAdminTab = () => {
-  return <>공지사항</>;
-};
-
-export default NoticeAdminTab;
