+++ src/pages/admin/NoticeManageTab/NoticeCreateTab.tsx
@@ -1,49 +1,33 @@
 import Input from '@components/Input';
 import RoundedButton from '@components/RoundedButton';
 import TextArea from '@components/TextArea';
-import { useMutation, useQuery } from '@tanstack/react-query';
-import { getNoticeDetail, patchNotice } from 'apis/notices';
+import { useMutation } from '@tanstack/react-query';
+import { postCreateNotice } from 'apis/notices';
+import useGoBack from 'hooks/useGoBack';
 import { useToast } from 'hooks/useToast';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
-interface Props {
-  noticeId: number;
-  onGoBack: () => void;
-}
-
-const NoticeEditSection = ({ noticeId, onGoBack }: Props) => {
+const NoticeCreateTab = () => {
+  const { goBack } = useGoBack();
   const toast = useToast();
+
   const [title, setTitle] = useState('');
   const [description, setDescription] = useState('');
 
-  const { data: notice, isError } = useQuery({
-    queryKey: ['noticeDetail', noticeId],
-    queryFn: () => getNoticeDetail(noticeId),
-  });
-
-  const editMutation = useMutation({
-    mutationFn: () => patchNotice(noticeId!, { title, description }),
+  const createMutation = useMutation({
+    mutationFn: () => postCreateNotice({ title, description }),
     onSuccess: () => {
       toast(`공지사항이 작성 되었어요.`, 'success');
-      onGoBack();
+      goBack();
     },
     onError: () => toast(`공지사항 작성에 실패했어요.`, 'error'),
   });
 
-  useEffect(() => {
-    if (notice) {
-      setTitle(notice.title || '');
-      setDescription(notice.description || '');
-    }
-  }, [notice]);
+  const handleSave = () => createMutation.mutate();
 
-  const handleSave = () => editMutation.mutate();
-  if (isError) {
-    return <div className="text-red-500">공지사항을 불러오는 도중 오류가 발생했습니다</div>;
-  }
   return (
     <section className="flex flex-col gap-8">
-      <h3 className="text-2xl font-bold">공지사항 수정</h3>
+      <h3 className="text-2xl font-bold">공지사항 생성</h3>
       <div className="grid grid-cols-[max-content_1fr] gap-x-8 gap-y-4">
         <label htmlFor="title" className="m-2">
           제목
@@ -62,7 +46,7 @@ const NoticeEditSection = ({ noticeId, onGoBack }: Props) => {
         />
       </div>
       <div className="flex justify-end gap-4">
-        <RoundedButton className="min-w-28" onClick={onGoBack}>
+        <RoundedButton className="min-w-28" onClick={goBack}>
           취소
         </RoundedButton>
         <RoundedButton className="min-w-28" onClick={handleSave}>
@@ -73,4 +57,4 @@ const NoticeEditSection = ({ noticeId, onGoBack }: Props) => {
   );
 };
 
-export default NoticeEditSection;
+export default NoticeCreateTab;
