+++ src/pages/admin/NoticeManageTab/NoticeEditTab.tsx
@@ -2,35 +2,48 @@ import Input from '@components/Input';
 import RoundedButton from '@components/RoundedButton';
 import TextArea from '@components/TextArea';
 import { useMutation, useQuery } from '@tanstack/react-query';
-import { getNoticeDetail, patchNotice, postCreateNotice } from 'apis/notices';
+import { getNoticeDetail, patchNotice } from 'apis/notices';
+import useGoBack from 'hooks/useGoBack';
 import { useToast } from 'hooks/useToast';
 import { useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
 
-interface Props {
-  noticeId?: number;
-  onGoBack: () => void;
-}
-
-const NoticeCreateSection = ({ noticeId, onGoBack }: Props) => {
+const NoticeEditTab = () => {
+  const { goBack } = useGoBack();
   const toast = useToast();
+  const noticeId = Number(useParams().noticeId);
 
   const [title, setTitle] = useState('');
   const [description, setDescription] = useState('');
 
-  const createMutation = useMutation({
-    mutationFn: () => postCreateNotice({ title, description }),
-    onSuccess: () => {
-      toast(`공지사항이 작성 되었어요.`, 'success');
-      onGoBack();
-    },
-    onError: (error) => toast(`공지사항 작성에 실패했어요.`, 'error'),
+  const { data: notice, isError } = useQuery({
+    queryKey: ['noticeDetail', noticeId],
+    queryFn: () => getNoticeDetail(noticeId),
   });
 
-  const handleSave = () => createMutation.mutate();
+  const editMutation = useMutation({
+    mutationFn: () => patchNotice(noticeId, { title, description }),
+    onSuccess: () => {
+      toast(`공지사항이 작성 되었어요.`, 'success');
+      goBack();
+    },
+    onError: () => toast(`공지사항 작성에 실패했어요.`, 'error'),
+  });
 
+  useEffect(() => {
+    if (notice) {
+      setTitle(notice.title || '');
+      setDescription(notice.description || '');
+    }
+  }, [notice]);
+
+  const handleSave = () => editMutation.mutate();
+  if (isError) {
+    return <div className="text-red-500">공지사항을 불러오는 도중 오류가 발생했습니다</div>;
+  }
   return (
     <section className="flex flex-col gap-8">
-      <h3 className="text-2xl font-bold">공지사항 생성</h3>
+      <h3 className="text-2xl font-bold">공지사항 수정</h3>
       <div className="grid grid-cols-[max-content_1fr] gap-x-8 gap-y-4">
         <label htmlFor="title" className="m-2">
           제목
@@ -49,7 +62,7 @@ const NoticeCreateSection = ({ noticeId, onGoBack }: Props) => {
         />
       </div>
       <div className="flex justify-end gap-4">
-        <RoundedButton className="min-w-28" onClick={onGoBack}>
+        <RoundedButton className="min-w-28" onClick={goBack}>
           취소
         </RoundedButton>
         <RoundedButton className="min-w-28" onClick={handleSave}>
@@ -60,4 +73,4 @@ const NoticeCreateSection = ({ noticeId, onGoBack }: Props) => {
   );
 };
 
-export default NoticeCreateSection;
+export default NoticeEditTab;
